#92 feat: 순환참조를 풀기 위한 ChatConsultService 생성
